fix(db): read NUMERIC aggregates as the type Postgres actually returns
`GET /ui-bridge/analytics/health-score` panicked on EVERY call, including
against an empty table, and was the only panicking route in a 105-route
read-only GET sweep:

    500 {"success":false,"error":"handler panicked:
         error retrieving column 2: error deserializing column 2"}

`task_run_automation.duration_ms` is `integer`, and PostgreSQL's
`AVG(integer)` returns **numeric**, which tokio-postgres will not
deserialize into `f64`. `query_one` over bare aggregates always returns
exactly one row, so the read at column 2 was reached unconditionally —
zero rows in the table did not save it, and neither did the `Option`,
which rescues a NULL and never a wrong type. The comment two lines above
documented the identical class being fixed for column 1 (a
`COALESCE(SUM(...))` NULL) and stopped there.

The sweep for siblings found 23 more sites. The rule that decides each
one is the ARGUMENT type, not the function name: `AVG` of any exact
numeric type is numeric; `SUM(int4)` is bigint but `SUM(int8)` is
numeric; a bare `100.0` is a numeric constant, so `bigint * 100.0` leaves
the integer domain; `ROUND(x, n)` is numeric; and `COALESCE` keeps
numeric unless one arm is already the preferred `double precision`.

NUMERIC read as f64/i64:

- tiered_info.rs `compute_automation_health_score` — `AVG(duration_ms)`
  (integer) into `Option<f64>`. PANICS ALWAYS. The reported defect.
- tiered_info.rs `get_performance_dashboard` — `AVG(tra.duration_ms)`,
  same column, same shape, second route. PANICS ALWAYS.
- meta_optimizer.rs `capture_snapshot` cost query — `COALESCE(AVG(
  run_cost), 0.0)` over `SUM(cost_cents)` (bigint) into `f64`. PANICS
  ALWAYS, so snapshot capture was unreachable.
- spec_experimentation.rs `build_model_profile_data` — `COALESCE(AVG(
  run_cost), 0)` over `SUM(cost_cents) / 100.0` into `f64`. PANICS
  ALWAYS. `.map(|r| r.get(0)).unwrap_or(0.0)` is no guard: `get` panics
  rather than returning an `Err` the `map` could swallow.
- graph_ops.rs `get_phase_stats` — all four aggregate columns at once:
  `AVG(duration_ms)`, `COALESCE(SUM(token_count), 0)` (both bigint) and
  `AVG(validation_errors_before/after)` (integer).
- meta_optimizer.rs architecture breakdown — `SUM(CASE ...) * 100.0 /
  COUNT(*)` into `f64`.
- meta_optimizer.rs `get_agent_aggregates_for_period` — `AVG(duration_ms)`
  (bigint) into `f64`.
- meta_optimizer.rs verification failures — `ROUND(100.0 * SUM / NULLIF(
  SUM), 1)` into `Option<f64>`.
- meta_optimizer.rs pipeline agent failures — `AVG(pat.duration_ms)`
  (bigint) into `Option<f64>`.

TIMESTAMPTZ read as String — the same panic, a different column type:

- reflection.rs `rebuild_architecture` — `MAX(created_at)`.
- meta_optimizer.rs recurring issues — `MAX(trf.detected_at)`.
- meta_optimizer.rs L1/L2 detail readers — five `created_at` columns
  selected without the `::TEXT` every other read site in that file
  spells out.

INTEGER read as i64:

- meta_optimizer.rs — `learning_outcomes.iterations`,
  `workflow_generation_feedback.rating`,
  `workflow_verification_phase_results.iteration`, and
  `meta_optimizer_runs.runs_analyzed` /`.recommendations_produced` at
  both list sites. The third read site of those last two takes them as
  `i32` and is correct, so they are cast per statement rather than
  per column.

Three write-side binds of PR #1238's own class that the earlier sweep
missed, each failing 100% of the time in BIND:

- reflection.rs `architecture_components.last_activity_at = $9` — bare
  `$n` inferred as timestamptz from the target column, bound from
  `Option<String>`.
- meta_optimizer.rs `complete_optimizer_run` — `completed_at = $3` the
  same way, plus `runs_analyzed`/`recommendations_produced` binding
  Rust `i64` into `integer` columns.
- tiered_info.rs `record_workflow_learning` —
  `learning_outcomes.created_at` from a `String`.

One SQL error the numeric fix would otherwise have kept invisible:
`get_agent_aggregates_for_period` compared
`pipeline_agent_traces.downstream_success` — a `boolean` — with `= 1`
and `= 0`. Postgres has no `boolean = integer` operator, so the whole
statement failed to parse and the function errored on every call.

Separately, the four `started_at::TEXT > $n` predicates in tiered_info.rs
are a live defect, not just an old spelling. They compared Postgres's own
rendering of a timestamptz (`2019-03-15 18:00:00+00`) against an RFC3339
string (`2019-03-15T12:00:00+00:00`) LEXICALLY. Both agree through the
date and diverge at the separator, where ' ' (0x20) sorts below 'T'
(0x54) — so every row landing on the cutoff's own calendar day was
dropped whatever its clock time, and `?days=0` excluded the whole of
today. They now compare real timestamps, with `$n::TEXT::TIMESTAMPTZ`
keeping the parameter typed as text so it still binds from a `String`
(PR #1238). The rendering also follows the session TimeZone, which the
RFC3339 parameter does not.

Sites deliberately left alone: every `SUM(CASE WHEN ... THEN 1 ELSE 0
END)` (int4 literals, so bigint, read as `i64`); every `COUNT(*)` read as
`i64`; every aggregate already carrying `::BIGINT` / `::float8` /
`::double precision`; `AVG` over columns that are already `double
precision`; the `COALESCE(ROUND(AVG(...)::numeric, 1), 0)` pair in
`get_optimizer_context`, whose column is never read at all; and the
whole of mcp/meta_optimizer_api.rs, which is inside a
`/* BEGIN_REMOVED ... END_REMOVED */` block and does not compile.

Adds four PG-backed regression tests that own a fixed slice of the
timeline rather than a config id, because the health-score query filters
on nothing but `started_at`. They cover the empty window (where the panic
needed no data), a populated window (asserting the mean is the real mean
and the score is 2/3), the cutoff-day window boundary, and the
performance-dashboard sibling. Neutering the `::float8` turns the first
two red with the original `error deserializing column 2`; neutering the
window predicate turns the boundary test red at 1 row instead of 2.

+++ src-tauri/src/database/pg/spec_experimentation.rs
@@ -810,7 +810,14 @@ impl PgDb {
 
         let avg_cost: f64 = conn
             .query_one(
-                r#"SELECT COALESCE(AVG(run_cost), 0) FROM (
+                // `cost_cents` is `bigint`, so `SUM` is numeric, `/ 100.0`
+                // keeps it numeric, `AVG(numeric)` is numeric and
+                // `COALESCE(numeric, 0)` still is. `f64` refuses numeric, and
+                // `.map(|r| r.get(0)).unwrap_or(0.0)` is no guard at all —
+                // `get` PANICS rather than returning an `Err` the `map` could
+                // swallow. `query_one` over a bare aggregate always returns a
+                // row, so this fired against an empty table too.
+                r#"SELECT COALESCE(AVG(run_cost), 0)::float8 FROM (
                        SELECT ptu.task_run_id, SUM(ptu.cost_cents) / 100.0 as run_cost
                        FROM phase_token_usage ptu
                        WHERE ptu.model_used = $1 AND ptu.created_at > $2
